+++ pydantic/errors.py
@@ -270,4 +270,4 @@ class PatternError(PydanticValueError):
 
 class DataclassTypeError(PydanticTypeError):
     code = 'dataclass'
-    msg_template = 'must be an instance of {class_name}'
+    msg_template = 'instance of {class_name}, tuple or dict expected'