+++ HISTORY.rst
@@ -12,6 +12,7 @@ v0.17.0 (unreleased)
 * fix schema generation for fields annotated as ``: dict``, #330 by @nkonin
 * support for passing Config class in dataclasses decorator, #276 by @jarekkar
   (**breaking change**: this supersedes the ``validate_assignment`` argument with ``config``)
+* support for nested dataclasses, #334 by @samuelcolvin
 
 v0.16.1 (2018-12-10)
 ....................